requirements.txt
flask~=1.1.2
flask-socketio
eventlet
python-dotenv~=0.15.0
spotipy~=2.17.1